+++ packages/registry-ui-duckui/src/toast/index.ts
@@ -1,5 +1,0 @@
-export * from './toast'
-export * from './toast.type'
-export * from './toast.hook'
-export * from './toast.constants'
-export * from './toast.lib'