+++ packages/registry-ui-duckui/src/button/CHANGELOG.md
@@ -11,3 +11,8 @@ description: Updates and improvements to the Button component.
 - **Border Props:** Introduced new button border properties to allow borders with hover effects.
 - **Hook Update:** Removed state from the `useDuckShortcut` hook, previously imported from `@ahmedayoub/shortcut`.
 - **Documentation:** Updated documentation to reflect the latest changes to the Button component.
+
+### March 30th, 2025
+
+- **new api:** removed all the old code and made new api for the button component.
+- **coposabel:** made the button composable with the new api.